+++ dir_windows.go
@@ -1,0 +1,16 @@
+// this file contain dir type definition
+
+// +build windows
+
+package main
+
+import (
+	"os"
+)
+
+func dirBlocks(info *file, fi os.FileInfo) {
+}
+
+func getOwnerGroupInfo(fi os.FileInfo) (o string, g string) {
+	return
+}
