+++ go.mod
@@ -4,6 +4,7 @@ go 1.15
 
 require (
 	github.com/go-git/go-git/v5 v5.1.0
+	github.com/mattn/go-colorable v0.1.7 // indirect
 	github.com/pborman/getopt/v2 v2.0.0
 	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a
 )
